CpuCheck, writing of ram
Added cpucheck, writing into ram

+++ Config.h
@@ -5,6 +5,6 @@
 
 const unsigned long long CANARY = 0xBADC0FFEE;
 const unsigned int Seed = 33;
-typedef int elem_t;
+typedef double elem_t;
 
 #endif //CONFIG_H
